+++ priv/static/js/brando.auth.js
@@ -10076,7 +10076,7 @@ require.register('textfit', function(exports,req,module){
 
     })(exports,require,module);
   });
-})();require.register("web/static/js/brando_auth/brando_auth", function(exports, require, module) {
+})();require.register("brando_auth", function(exports, require, module) {
 "use strict";
 
 var _jquery = require("jquery");
@@ -10104,5 +10104,5 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 });
 });
 
-require('brando_auth/brando_auth');
+require('brando_auth');
 //# sourceMappingURL=brando.auth.js.map