+++ priv/static/js/brando.js
@@ -12854,6 +12854,16 @@ var _dropzone = require("dropzone");
 
 var _dropzone2 = _interopRequireDefault(_dropzone);
 
+require("./brando/extensions/dropdown");
+
+require("./brando/extensions/searcher");
+
+require("./brando/extensions/slugit");
+
+require("./brando/extensions/sparkline");
+
+require("./brando/extensions/tags_input");
+
 var _accordion = require("./brando/components/accordion");
 
 var _autoslug = require("./brando/components/autoslug");
@@ -12907,16 +12917,6 @@ var _utils = require("./brando/components/utils");
 var _ws = require("./brando/components/ws");
 
 var _ws2 = _interopRequireDefault(_ws);
-
-require("./brando/extensions/dropdown");
-
-require("./brando/extensions/searcher");
-
-require("./brando/extensions/slugit");
-
-require("./brando/extensions/sparkline");
-
-require("./brando/extensions/tags_input");
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
@@ -13175,8 +13175,6 @@ var _accordion = require("./accordion");
 
 var _utils = require("./utils");
 
-var _utils2 = _interopRequireDefault(_utils);
-
 var _vex_brando = require("./vex_brando");
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
@@ -13281,7 +13279,7 @@ var Images = function () {
                 headers: { Accept: "application/json; charset=utf-8" },
                 type: "POST",
                 data: data,
-                url: _utils2.default.addToPathName('set-properties')
+                url: _utils.Utils.addToPathName('set-properties')
             }).done(_jquery2.default.proxy(function (data) {
                 /**
                  * Callback after confirming.
@@ -13339,7 +13337,7 @@ var Images = function () {
                             _jquery2.default.ajax({
                                 headers: { Accept: "application/json; charset=utf-8" },
                                 type: "POST",
-                                url: _utils2.default.addToPathName('delete-selected-images'),
+                                url: _utils.Utils.addToPathName('delete-selected-images'),
                                 data: { ids: imagePool },
                                 success: that.deleteSuccess
                             });
@@ -15193,9 +15191,9 @@ _jquery2.default.fn.dropdown.Constructor.prototype.change = function (e) {
 require.register("brando/extensions/searcher", function(exports, require, module) {
 "use strict";
 
-var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol" ? function (obj) { return typeof obj; } : function (obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol ? "symbol" : typeof obj; }; /*! $ Searcher Plugin - v0.1.0 - 2014-08-18
+var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol" ? function (obj) { return typeof obj; } : function (obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol ? "symbol" : typeof obj; }; /*! jQuery Searcher Plugin - v0.3.0 - 2016-01-29
                                                                                                                                                                                                                                                    * https://github.com/lloiser/jquery-searcher/
-                                                                                                                                                                                                                                                   * Copyright (c) 2014 Lukas Beranek; Licensed MIT
+                                                                                                                                                                                                                                                   * Copyright (c) 2016 Lukas Beranek; Licensed MIT
                                                                                                                                                                                                                                                   */
 
 var _jquery = require("jquery");
@@ -15245,6 +15243,30 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 					toggle(this, true);
 				});
 			},
+			filter: function filter(value) {
+				this._lastValue = value;
+
+				var options = this.options,
+				    textSelector = options.textSelector,
+				    toggle = options.toggle || defaults.toggle;
+
+				// build the regular expression for searching
+				var flags = "gm" + (!options.caseSensitive ? "i" : "");
+				var regex = new RegExp("(" + escapeRegExp(value) + ")", flags);
+
+				this._$element.find(options.itemSelector).each(function eachItem() {
+					var $item = $(this),
+					    $textElements = textSelector ? $item.find(textSelector) : $item,
+					    itemContainsText = false;
+
+					$textElements = $textElements.each(function eachTextElement() {
+						itemContainsText = itemContainsText || !!$(this).text().match(regex);
+						return !itemContainsText; // stop if at least one text element contains the text
+					});
+
+					toggle(this, itemContainsText);
+				});
+			},
 			_create: function _create() {
 				var options = this.options;
 
@@ -15256,7 +15278,7 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 				this._$input = $(options.inputSelector).bind(eventNames, this._fn);
 
 				// remember the last entered value
-				this._lastValue = "";
+				this._lastValue = null;
 
 				// call the toggle with true for all items on startup
 				var toggle = options.toggle || defaults.toggle;
@@ -15265,30 +15287,10 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 				});
 			},
 			_onValueChange: function _onValueChange() {
-				var options = this.options,
-				    textSelector = options.textSelector,
-				    toggle = options.toggle || defaults.toggle;
+				var value = this._$input.val();
+				if (value === this._lastValue) return; // nothing has changed
 
-				// build the regular expression for searching
-				var flags = "gm" + (!options.caseSensitive ? "i" : "");
-				var value = new RegExp("(" + escapeRegExp(this._$input.val()) + ")", flags);
-
-				if (value.toString() === this._lastValue) return; // nothing has changed
-
-				this._lastValue = value.toString();
-
-				this._$element.find(options.itemSelector).each(function eachItem() {
-					var $item = $(this),
-					    $textElements = textSelector ? $item.find(textSelector) : $item,
-					    itemContainsText = false;
-
-					$textElements = $textElements.each(function eachTextElement() {
-						itemContainsText = itemContainsText || !!$(this).text().match(value);
-						return !itemContainsText; // stop if at least one text element contains the text
-					});
-
-					toggle(this, itemContainsText);
-				});
+				this.filter(value);
 			}
 		};
 
@@ -15298,23 +15300,25 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 		}
 
 		$.fn[pluginName] = function pluginHandler(options) {
+			var args = Array.prototype.slice.call(arguments, 1);
 			return this.each(function () {
 				var searcher = $.data(this, dataKey);
-				if (searcher && options === "dispose") {
-					searcher.dispose();
-					$.removeData(this, dataKey);
+				var t = typeof options === "undefined" ? "undefined" : _typeof(options);
+				if (t === "string" && searcher) {
+					searcher[options].apply(searcher, args);
+					if (options === "dispose") $.removeData(this, dataKey);
+				} else if (t === "object") {
+					if (!searcher)
+						// create a new searcher
+						$.data(this, dataKey, new Searcher(this, options));else
+						// update the options of the existing
+						$.extend(searcher.options, options);
 				}
-				// update the options of the existing
-				else if (searcher) $.extend(searcher.options, options);
-					// create a new searcher
-					else if ((typeof options === "undefined" ? "undefined" : _typeof(options)) === "object") $.data(this, dataKey, new Searcher(this, options));
 			});
 		};
 	}
-	// node/CommonJS style (for Browserify)
-	if ((typeof exports === "undefined" ? "undefined" : _typeof(exports)) === "object") module.exports = factory;
-	// browser
-	else factory(_jquery2.default);
+
+	factory(_jquery2.default);
 }).call(undefined);
 });
 
@@ -16021,7 +16025,7 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
                       */
 });
 
-require.register("web/static/js/brando_auth/brando_auth", function(exports, require, module) {
+require.register("brando_auth", function(exports, require, module) {
 "use strict";
 
 var _jquery = require("jquery");